feat: add use-home hook
